Converge AppService and MediaControlService into BaeKit
The two apps carried hand-forked copies of the root application object and
the media-control bridge that had drifted apart while doing the same job.
Fold the shared halves into BaeKit: a shared AppService base class each app
subclasses with its own construction and wiring, and one MediaControlService
whose platform-specific session handling lives in os-gated extensions. This
removes the duplicated projection registration and now-playing/scrub/artwork
logic, and gives macOS the scrubber management (enabled only while a track
duration is known) and iOS the bounded artwork decode that only one fork
previously had.
